refactor: centralize route flag logic and shared utility helpers into lib/route-utils.ts and lib/utils.ts
