+++ frontend/src/app/assess/[projectId]/[domain]/[practice]/page.tsx
@@ -23,23 +23,7 @@ import { buildAssessmentAnswerKey } from "../../../../../lib/assessmentValidatio
 import { useOptionalAssessmentContext } from "../../../../../contexts/AssessmentContext";
 import { Breadcrumb } from "@/components/shared/Breadcrumb";
 import { isPremiumStatus } from "@/lib/constants";
-
-const getDomainIcon = (title: string) => {
-    const t = title.toLowerCase();
-    if (t.includes("govern") || t.includes("responsib")) {
-        return IconShieldLock;
-    }
-    if (t.includes("data") || t.includes("privacy")) {
-        return IconDatabase;
-    }
-    if (t.includes("design") || t.includes("develop") || t.includes("model")) {
-        return IconCpu;
-    }
-    if (t.includes("security") || t.includes("protection")) {
-        return IconLock;
-    }
-    return IconFolder;
-};
+import { getDomainIcon } from "@/lib/utils";
 
 const getLevelColor = (level: string | number) => {
     const lvl = String(level);
